working on validations

+++ checkapp/views.py
@@ -16,12 +16,34 @@ def index(request):
 
 def create_checklist(request):
 
+    if 'email' not in request.session:
+        messages.error(request, 'Please log in to view this page.')
+        return redirect('/')
+
+    errors = Checklist.objects.basic_validator(request.POST)
+
+    if errors:
+        for k, v in errors.items():
+            messages.error(request, v)
+        return redirect('/checkapp')
+
     checklist = Checklist.objects.create(
         title = request.POST['title'],
         user = User.objects.get(id = request.session['user_id'])
     )
     return redirect('/checkapp')
 def create_item(request,checklist_id):
+
+    if 'email' not in request.session:
+        messages.error(request, 'Please log in to view this page.')
+        return redirect('/')
+
+    errors = Item.objects.basic_validator(request.POST)
+
+    if errors:
+        for k, v in errors.items():
+            messages.error(request, v)
+        return redirect('/checkapp')
 
     this_item = Item.objects.create(
         name = request.POST['name']
@@ -31,7 +53,21 @@ def create_item(request,checklist_id):
     
     return JsonResponse(this_item.return_JSON())
 def destroy_item(request,item_id):
+
+    if 'email' not in request.session:
+        messages.error(request, 'Please log in to view this page.')
+        return redirect('/')
+
     this_item = Item.objects.get(id = item_id)
     this_item.delete()
     context = {}
     return JsonResponse(context)
+def destroy_checklist(request, checklist_id):
+
+    if 'email' not in request.session:
+        messages.error(request, 'Please log in to view this page.')
+        return redirect('/')
+    
+    this_checklist = Checklist.objects.get(id = checklist_id)
+    this_checklist.delete()
+    return redirect('/checkapp')
